# internal/swagger/swagger.yaml
basePath: /
definitions:
...
  controller.OAuthURLSuccessResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      url:
        type: string
    type: object
...
  controller.SimpleResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Swagger documentation for Tinyauth's API.
  license:
    name: AGPL-3.0
    url: https://github.com/tinyauthapp/tinyauth/blob/main/LICENSE
  title: Tinyauth API
  version: development
paths:
...
  /api/oauth/callback/{id}:
    get:
      description: Callback URL for OAuth providers
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      - description: State
        in: query
        name: code
        required: true
        type: string
      - description: Code
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
      summary: OAuth Callback
      tags:
      - oauth
  /api/oauth/url/{id}:
    get:
      description: Get an OAuth URL for the specified provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      - description: Login for
        in: query
        name: login_for
        type: string
      - description: OpenID Connect Ticket
        in: query
        name: oidc_ticket
        type: string
      - description: OpenID Connect Scope
        in: query
        name: oidc_scope
        type: string
      - description: OpenID Connect Name
        in: query
        name: oidc_name
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.OAuthURLSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: OAuth URL
      tags:
      - oauth
...
swagger: "2.0"